requirements.txt
alabaster==0.7.12
anytree==2.8.0
appdirs==1.4.4
attrs==19.3.0
Babel==2.8.0
beautifulsoup4==4.9.1
bullet==2.1.0
certifi==2020.6.20
chardet==3.0.4
cliff==3.3.0
cmd2==1.1.0
colorama==0.4.3
configobj==5.0.6
configparser==5.0.0
distlib==0.3.0
docutils==0.16
filelock==3.0.12
gnupg==2.3.1
humanfriendly==8.2
idna==2.9
imagesize==1.2.0
Jinja2==2.11.2
lxml==4.5.1
MarkupSafe==1.1.1
numpy==1.19.0
packaging==20.4
pbr==5.4.5
pexpect==4.8.0
pluggy==0.13.1
prettytable==0.7.2
property-manager==3.0
psutil==5.7.0
ptyprocess==0.6.0
py==1.9.0
pycrypto==2.6.1
pyflakes==2.2.0
Pygments==2.6.1
pyparsing==2.4.7
pyperclip==1.8.0
python-secrets==20.4.2106
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
snowballstemmer==2.0.0
soupsieve==2.0.1
Sphinx==3.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
stevedore==2.0.1
toml==0.10.1
tox==3.15.2
urllib3==1.25.9
verboselogs==1.7
virtualenv==20.0.25
wcwidth==0.2.5
xkcdpass==1.17.3
yamlreader==3.0.4
